python(test): refactor test.py, move test code outside from test.py
